requirements.txt
# Libraries for code interpretation and execution
litellm>=1.83.0
litellm[proxy]
Pygments 
pyreadline 
python-dotenv 
pytest

# Libraries for making HTTP requests
Requests 
rich

# Libraries for data analysis
pandas
matplotlib
plotly

# Libraries for standard libries 
stdlib_list

# Legacy libraries
legacy-cgi